backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.45
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
pydantic==2.12.5
python-multipart==0.0.20
httptools==0.7.1
email-validator==2.3.0
psycopg2-binary==2.9.9
bcrypt==5.0.0
stripe
httpx
slowapi
